+++ src/main/java/de/alpharogroup/db/entity/uniqueable/UniqueableEntity.java
@@ -39,6 +39,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
+import lombok.experimental.SuperBuilder;
 
 /**
  * The class {@link UniqueableEntity} holds an UUID as primary key.
@@ -53,6 +54,7 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @FieldDefaults(level = AccessLevel.PRIVATE)
+@SuperBuilder
 public abstract class UniqueableEntity<PK extends Serializable>
 	implements
 		Serializable,