+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 ## Change log
 ----------------------
 
-Version 1.6-SNAPSHOT
+Version 1.7
+-------------
+
+ADDED:
+ 
+- new name entities that extends from the UUIDEntity
+CHANGED:
+
+- entities UniqueableEntity and UUIDEntity are with the lombok annotation 
+SuperBuilder annotated
+
+Version 1.6
 -------------
 
 ADDED: